Split storage/__internal__.py into multiple files.
