+++ azure-servicemanagement-legacy/azure/servicemanagement/_http/__init__.py
@@ -1,4 +1,4 @@
-#-------------------------------------------------------------------------
+﻿#-------------------------------------------------------------------------
 # Copyright (c) Microsoft.  All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,9 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #--------------------------------------------------------------------------
-
-HTTP_RESPONSE_NO_CONTENT = 204
-
 
 class HTTPError(Exception):
