+++ azure-storage/azure/storage/auth.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #--------------------------------------------------------------------------
-from ._internal import (
+from ._common_conversion import (
     _sign_string,
 )
