+++ azure-storage/azure/storage/blobservice.py
@@ -15,29 +15,30 @@
 from azure.common import (
     WindowsAzureError,
 )
-from ._internal import (
-    BLOB_SERVICE_HOST_BASE,
-    DEFAULT_HTTP_TIMEOUT,
-    DEV_BLOB_HOST,
-    _ERROR_VALUE_NEGATIVE,
-    _ERROR_PAGE_BLOB_SIZE_ALIGNMENT,
-    _convert_class_to_xml,
+from ._common_error import (
     _dont_fail_not_exist,
     _dont_fail_on_exist,
+    _validate_not_none,
+    _validate_type_bytes,
+    _ERROR_VALUE_NEGATIVE,
+    _ERROR_PAGE_BLOB_SIZE_ALIGNMENT,
+)
+from ._common_conversion import (
     _encode_base64,
+    _int_or_none,
+    _str,
+    _str_or_none,
+)
+from ._common_serialization import (
+    _convert_class_to_xml,
     _get_request_body,
     _get_request_body_bytes_only,
-    _int_or_none,
     _parse_response_for_dict,
     _parse_response_for_dict_filter,
     _parse_response_for_dict_prefix,
-    _str,
-    _str_or_none,
     _update_request_uri_query_local_storage,
-    _validate_type_bytes,
-    _validate_not_none,
     _ETreeXmlToObject,
-    )
+)
 from ._http import HTTPRequest
 from ._chunking import (
     _BlockBlobChunkUploader,
@@ -62,6 +63,9 @@ from .connection import (
     StorageConnectionParameters,
 )
 from .constants import (
+    BLOB_SERVICE_HOST_BASE,
+    DEFAULT_HTTP_TIMEOUT,
+    DEV_BLOB_HOST,
     X_MS_VERSION,
 )
 from ._serialization import (
