+++ azure-storage/azure/storage/models.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #--------------------------------------------------------------------------
-from ._internal import (
+from ._common_models import (
     WindowsAzureData,
     _list_of,
 )
